requirements.txt
langchain
langchain-core
langchain-community
pypdf
pymupdf
sentence-transformers
faiss-cpu
chromadb
langchain-groq
python-dotenv
typesense
langchain_openai
langgraph
wikipedia
arxiv
langchain-tavily
langchain-mcp-adapters
fastmcp
mcp